Trove add support for log retrieval
Add new log tab table to instance details to display the list of logs
that can be retrieved.  The table has row options to view the log
in a separate panel, publish new log data and stop collection of
log data.

The log data is displayed in a new panel.  In this panel the number
of lines to display can be changed.  Also checking the publish
checkbox and refreshing the log data will publish any new log data.
There are three buttons that perform the following actions:
* view the entire log in a separate window
* download the entire log
* return to the log list panel

The log code is placed in a subdirectory of database to keep the
code more manageable.

Added the supporting log commands to the api.

Change-Id: I2252f6f6f67946f62939472b8b88883b87e35982
Closes-Bug: #1548514