Add support for instance datastore-flavors
Update the launch wizard to use datastore-flavors instead of
flavor-list.

When selecting a datastore the flavor list will change to display only
flavors for the selected datastore.

If the datastore flavor is not configured then the default flavor
list will be used.

Updated unit tests to use the datastore flavors.

Change-Id: I91c8affcde9d410050f6a2dc8448af41638ba8a3
Closes-Bug: #1549367