Get all instances to populate master pulldown
The call to gather all instances to populate the master pulldown in
the launch database instance wizard will never exceed the default
pagination size because the call is only made once and never checks
if there is more data.

Looped the call to obtain instances until there is no more data to
retrieve.

Added a unit test.

Change-Id: I3574c176d333b46f07b1551e23c8e8298ce7f0e3
Closes-Bug: #1549937