Add missing replication functionality
Add Horizon support for ejecting replica sources, promoting replicas
and specifying the replica count when launching instances based on
a master instance.

Added a promote row action to the instance table that is available
only when the instance is active and is a replica.
Added a confirmation dialog that displays information on the current
replica and replica source.

Added an eject row action to the instances table that is only
visible when the instance is a replica source.

Added an integer field to specify the number of replicas to start
on the advanced tab in the Launch instance dialog.  This field
appears only when a replica source is selected and has a minimum
value of 1.

Modified the trove create instance api to include the replica count
and new apis for eject and promote.

Change-Id: Id46d039abd38b2d2389d19a3e741b8a13f552d2d
Closes-Bug: #1551807