Improved handling of Swift panel detection
The initial handling of the Swift panel detection can be improved
by more closely following the way the detection is done in the
Swift Horizon code.  Namely, if the setting is not angular then the
original default Django panel will be enabled.

Also added code to detect if the setting is missing which would
also result in the default Django panel being the default.

Change-Id: I91567b1f673bf89af46a572144acde93f4e1b316