Add support for uppercase fieldname to cluster
The datastore flavor field was not properly implemented because
not every datastore flavor had it's own flavor pulldown.
Properly implemented a datastore pulldown for every flavor that
is only visible when the corresponding datastore value is selected.

Also added code to protect against uppercase letters in the
datastore or datastore version name.

Change-Id: Ie5f45be4f8f5925cc3b34f0b5b32a02d459396fc
Closes-Bug: DBAAS-1546