Add support for upgrade instance
Added a row menu action for upgrade instance only if there are
multiple datastore versions for the instance datastore.

The action will bring up a dialog showing the current version
and a dropdown listing all the datastore versions (with the
current version filterd out of the list).

Added a new trove api for instance_upgrade.

Change-Id: I0257d1650da870931f4324554f1039f7582af74f
Closes-Bug: DBAAS-1203