Use MySQL qualified names for user actions
Added code in db_capability to provide a fully qualified MySQL name
if the datastore is MySQL (or a variant like MariaDB).
User actions will call the db_capability code to get the name as it
exists now or the new fully qualified MySQL name as determined by
the datastore.

Change-Id: I1d7cd11f74ed35f7c63125e5f8a77003983dfc5c
Closes-Bug: DBAAS-1439