Add region support for instances and clusters
Modified the launch cluster form and the add instance during grow
form to display the list of available regions.

Modified the cluster manager to account for the region.

Added the region value to the api calls to launch instance/cluster
and grow add instance.  Add the trove api call to use keystone to
obtain the region list.

Updated unit tests.

Change-Id: I5492476a3ef49120516a871c50121b4067d1ca84
Closes-Bug: DBAAS-1743