Restrict launch fields when restoring from backup
When restoring from backup the launch wizard contains all options
and performs no customizations based on the information contained
in the backup.

This patch adds support for restricting the launch instance
datastore field to the datastore and datastore version that the
backup is relevant to.

It also populates the restored backup as the only option in the
advanced step backup field.

Added a unit test to validate the fields were populated correctly.

Change-Id: Iabfefd34b5f7a24699277104fa92ac77b2c74fe3
Closes-Bug: DBAAS-1848