Disable admin columns in datastore version table.
The admin columns will cause the panel to throw and exception if
the user is not admin. The code to govern the admin columns is
stuck in US horizon and until that gets merged these columns
should be removed until the US policy patch is merged as there is
no other way to determine the user in the table.

Change-Id: I9c0f89cc7934b3b67d5d45b21adfbf5aa0e59e2e
Closes-Bug: DBAAS-1865