Final fixes for region support
Updated api parameter names to match server implementation.

Use the more widely available openstack_auth library to retrieve
the list of regions.

Added code to display "No regions found" if the region list is
empty.  Also if there is only one region then it will be the only
value in the dropdown.  If there are more than one then the
"Default region" will be the default choice.

Added code to enable the region support on the cluster launch
dialog and the add instance during cluster grow dialog.

Change-Id: If6d067cb0f7a7b75316752c77a88ff031cf78699
Closes-Bug: DBAAS-1743