Add view of replication topology
Added a new panel for displaying the replication topology.

The panel will display the list of master instances from which
the replicas can be view by selecting the view replicas action.

The view replicas action will display a new page with the replicas
for the selected master.

Related replication actions will be available on the master instance
table and the replica table.

Also refactored some code to enable reuse of existing code instead
of having it duplicated in multiple places.

Change-Id: I1da71f057d589e1745336fdfcdb10683eec0ee4e
Closes-Bug: DBAAS-1134